-Refactored the latency and time datatype everywhere from float to double -Added the possibility to low-pass filter the found latencies -Manually tested with teensy, added the results to the project

+++ SignalSync/src/be/signalsync/core/SyncEventListener.java
@@ -10,5 +10,5 @@ import java.util.List;
  *
  */
 public interface SyncEventListener {
-	void onSyncEvent(List<Float> data);
+	void onSyncEvent(List<Double> data);
 }
