+++ SignalSync/src/be/signalsync/syncstrategy/FingerprintSyncStrategy.java
@@ -95,18 +95,18 @@ public class FingerprintSyncStrategy extends SyncStrategy {
 	 *         match found, NaN is added to the list.
 	 */
 	@Override
-	public List<Float> findLatencies(List<float[]> slices) {
+	public List<Double> findLatencies(List<float[]> slices) {
 		if(slices.isEmpty()) {
 			throw new IllegalArgumentException("The slices list can not be empty.");
 		}
-		final float fftHopSizesS = stepSize / (float) sampleRate;
-		final List<Float> latencies = new ArrayList<>();
+		final double fftHopSizesS = stepSize / (double) sampleRate;
+		final List<Double> latencies = new ArrayList<>();
 		for (final int[] timing : synchronize(slices)) {
 			if (timing.length > 0) {
 				// Calculating the time difference from the time index
 				latencies.add(timing[0] * fftHopSizesS - timing[1] * fftHopSizesS);
 			} else {
-				latencies.add(Float.NaN);
+				latencies.add(Double.NaN);
 			}
 		}
 		return latencies;
