+++ SignalSync/src/be/signalsync/test/RecordedTest.java
@@ -95,7 +95,7 @@ public class RecordedTest {
 	@Test
 	public void testCrossCovariance() {
 		Config.set(Key.LATENCY_ALGORITHM, "crosscovariance");
-		final List<Float> latencies = crossCovarianceStrategy.findLatencies(streams);
+		final List<Double> latencies = crossCovarianceStrategy.findLatencies(streams);
 		Assert.assertEquals("The result should contain 1 latency", 1, latencies.size());
 		Assert.assertEquals(String.format("Crosscovariance failed when latency: %.4f, type: %s", latency, type), 
 				latency, latencies.get(0), 0.002);
@@ -104,7 +104,7 @@ public class RecordedTest {
 	@Test
 	public void testFingerprint() {
 		Config.set(Key.LATENCY_ALGORITHM, "fingerprint");
-		final List<Float> latencies = fingerprintSyncStrategy.findLatencies(streams);
+		final List<Double> latencies = fingerprintSyncStrategy.findLatencies(streams);
 		Assert.assertEquals("The result should contain 1 latency", 1, latencies.size());
 		Assert.assertEquals(String.format("Fingerprinting failed when latency: %.4f, type: %s", latency, type), 
 				latency, latencies.get(0), 0.032);
